+++ thoth-errors/src/database_errors.rs
@@ -1,5 +1,5 @@
-use phf::phf_map;
-use phf::Map;
+use phf::{phf_map, Map};
+use std::borrow::Cow;
 
 use crate::ThothError;
 
@@ -173,7 +173,7 @@ impl From<diesel::result::Error> for ThothError {
             Error::DatabaseError(_kind, info) => {
                 if let Some(constraint_name) = info.constraint_name() {
                     if let Some(error) = DATABASE_CONSTRAINT_ERRORS.get(constraint_name) {
-                        return ThothError::DatabaseConstraintError(error);
+                        return ThothError::DatabaseConstraintError(Cow::Borrowed(error));
                     }
                 }
                 ThothError::DatabaseError(info.message().to_string())
@@ -251,9 +251,9 @@ mod tests {
                 DatabaseErrorKind::UniqueViolation,
                 error_information
             )),
-            ThothError::DatabaseConstraintError(
+            ThothError::DatabaseConstraintError(Cow::Borrowed(
                 "A contribution with this ordinal number already exists."
-            )
+            ))
         )
     }
     #[test]